Fix the list functions's bug and update the delete functions, update some words from English to Turkish, block to list admin user.

+++ backend/modules/books.py
@@ -14,7 +14,7 @@ class Book:
 
         try:
 
-            if bookName.replace(" ","") == "" or writer.replace(" ","") == "" or publisher.replace(" ","") ==  "" or pageCount == 0 or pageCount < 0:
+            if not bookName.strip() or not writer.strip() or not publisher.strip() or pageCount <= 0:
                 return {"success":False,"message":"Lütfen boş bırakmayın!"}
             
             self.cursor.execute("SELECT * FROM categories WHERE categoryName = %s",(category,))
@@ -38,13 +38,14 @@ class Book:
             if int(pageCount) > 99999:
                 return {"success":False,"message":"Sayfa sayısı 99999'dan fazla olamaz!"}
                                     
-            self.cursor.execute("INSERT INTO books (bookName,writer,category,publisher,pageCount,whoAdded) VALUES (%s,%s,%s,%s,%s,%s)",(bookName,writer,category,publisher,pageCount,activeUserName))
+            self.cursor.execute("INSERT INTO books (bookName,writer,category,publisher,pageCount,whoAdded) VALUES (%s,%s,%s,%s,%s,%s)",(bookName.strip(),writer.strip(),category.strip(),publisher.strip(),pageCount,activeUserName))
             self.conn.commit()
                                         
             return {"success":True,"message":"Kitap eklendi."}
 
         except Exception as e:
 
+            self.conn.rollback()
             writeLog(config.BOOKS_LOG_PATH,type(e).__name__,str(e))
             return {"success":False,"message":"Bir hata oluştu!"}
 
@@ -53,7 +54,7 @@ class Book:
 
         try:
 
-            if id == 0 or id < 0:
+            if id <= 0:
                 return {"success":False,"message":"Lütfen boş bırakmayın!"}
             
             self.cursor.execute("SELECT * FROM books WHERE id = %s",(id,))
@@ -62,6 +63,12 @@ class Book:
             if result is None:
                 return {"success":False,"message":"Kitap bulunamadı!"}
 
+            self.cursor.execute("SELECT * FROM books WHERE id = %s and isTaken = 'Alındı'",(id,))
+            result = self.cursor.fetchone()
+
+            if result is not None:
+                return {"success":False,"message":"Kitap ödünç verildiği için silinemez!"}
+            
             self.cursor.execute("DELETE FROM books WHERE id = %s",(id,))
             self.conn.commit()
 
@@ -69,8 +76,8 @@ class Book:
 
         except Exception as e:
 
+            self.conn.rollback()
             writeLog(config.BOOKS_LOG_PATH,type(e).__name__,str(e))
-            
             return {"success":False,"message":"Bir hata oluştu!"}
     
 
@@ -80,35 +87,30 @@ class Book:
 
             filterList = ["id","bookName","writer","category","publisher","pageCount","isTaken","whoAdded"]
 
-            if limit == 0 or limit < 0 or pageNumber == 0 or pageNumber < 0:
+            if limit <= 0 or pageNumber <= 0:
                 return {"success":False,"message":"Lutfen boş bırakmayın!"}
             
-            if limit > 50:
-                return {"success":False,"message":"Limit en fazla 50 olabilir!"}
+            if limit > 10:
+                return {"success":False,"message":"Sayfaya düşen satır sayısı en fazla 10 olabilir!"}
             
-            IDs,names,writers,categories,publishers,pageCounts,isTakens,whoAddeds = [],[],[],[],[],[],[],[]
-            offset = ((pageNumber - 1) * limit)
+            books = []
 
             self.cursor.execute("SELECT COUNT(*) FROM books")
             totalBooks = self.cursor.fetchone()[0]
+
             if totalBooks is not None:
                 pageCount = totalBooks // limit
                 if totalBooks % limit != 0:
                     pageCount += 1
 
+            offset = ((pageNumber - 1) * limit)
+
             def add(rV):
-                IDs.append(rV[0])
-                names.append(rV[1])
-                writers.append(rV[2])
-                publishers.append(rV[3])
-                pageCounts.append(rV[4])
-                categories.append(rV[5])
-                isTakens.append(rV[6])
-                whoAddeds.append(rV[7])
+                books.append({"ID":rV[0],"name":rV[1],"writer":rV[2],"publisher":rV[3],"pageCount":rV[4],"category":rV[5],"isTaken":rV[6],"whoAdded":rV[7]})
 
             if isWithFilter:
 
-                if filterValue.replace(" ","") == "" or filterType.replace(" ","") == "":
+                if not filterValue.strip() or not filterType.strip():
                     return {"success":False,"message":"Lütfen boş bırakmayın!"}
                         
                 if filterType != "pageCount" and filterType != "id":
@@ -119,12 +121,14 @@ class Book:
                     return {"success":False,"message":"Lütfen geçerli parametre giriniz!"}
                               
                 if filterType != "pageCount" and filterType != "id":
-                    newFilterValue = f"%{filterValue}%"
+                    newFilterValue = f"%{filterValue.strip()}%"
                     self.cursor.execute(f"SELECT COUNT(*) FROM books WHERE {filterType} LIKE %s", (newFilterValue,))
                     totalBooks = self.cursor.fetchone()[0]
-                    pageCount = totalBooks // limit
-                    if totalBooks % limit != 0:
-                        pageCount += 1
+
+                    if totalBooks is not None:
+                        pageCount = totalBooks // limit
+                        if totalBooks % limit != 0:
+                            pageCount += 1
 
                     self.cursor.execute(f"SELECT * FROM books WHERE {filterType} LIKE %s LIMIT %s OFFSET %s", (newFilterValue,limit, offset))
                     result = self.cursor.fetchall()
@@ -133,45 +137,36 @@ class Book:
                                 
                     self.cursor.execute(f"SELECT COUNT(*) FROM books WHERE {filterType} = %s", (int(filterValue),))
                     totalBooks = self.cursor.fetchone()[0]
-                    pageCount = totalBooks // limit
-                    if totalBooks % limit != 0:
-                        pageCount += 1
+
+                    if totalBooks is not None:
+                        pageCount = totalBooks // limit
+                        if totalBooks % limit != 0:
+                            pageCount += 1
 
                     self.cursor.execute(f"SELECT * FROM books WHERE {filterType} = %s LIMIT %s OFFSET %s", (int(filterValue),limit, offset))
                     result = self.cursor.fetchall()
-
-                    if result:
-                        for r in result:
-                            add(rV=r)  
                     
-            elif isWithFilter == False:
+            else:
 
                 self.cursor.execute("SELECT * FROM books LIMIT %s OFFSET %s", (limit, offset))
-                result2 = self.cursor.fetchall()
+                result = self.cursor.fetchall()
 
-                if result2: 
-                    for r2 in result2:
-                        add(rV=r2)
+            if result: 
+                for r2 in result:
+                    add(rV=r2)
                     
-                if len(IDs) == 0:
-                    return {"success":False,"message":"Sonuç bulunamadı!"}
+            if len(books) == 0:
+                return {"success":False,"message":"Sonuç bulunamadı!"}
                 
-                return {
-                    "success":True,
-                    "message":"Kitaplar listelendi.",
-                    "data":{
-                        "totalBooks":totalBooks,
-                        "ids":IDs,
-                        "names":names,
-                        "writers":writers,
-                        "categories":categories,
-                        "publishers":publishers,
-                        "pageCounts":pageCounts,
-                        "isTakens":isTakens,
-                        "whoAddeds":whoAddeds,
-                        "pageCount":pageCount
-                    }
+            return {
+                "success":True,
+                "message":"Kitaplar listelendi.",
+                "data":{
+                    "totalBooks":totalBooks,
+                    "books":books,
+                    "pageCount":pageCount
                 }
+            }
 
                             
 
@@ -184,7 +179,7 @@ class Book:
 
         try:
 
-            if id == 0 or id < 0 or bookName.replace(" ","") == "" or writer.replace(" ","") == "" or publisher.replace(" ","") == "" or pageCount == 0 or category.replace(" ","") == "" or pageCount < 0:
+            if id <= 0 or not bookName.strip() or not writer.strip() or not publisher.strip() or pageCount <= 0 or not category.strip():
                 return {"success":False,"message":"Lütfen boş bırakmayın!"}
             
             self.cursor.execute("SELECT * FROM books WHERE id = %s",(id,))
@@ -214,12 +209,13 @@ class Book:
             if int(pageCount) > 99999:
                 return {"success":False,"message":"Sayfa sayısı 99999'dan fazla olamaz!"}
 
-            self.cursor.execute("UPDATE books SET bookName = %s, writer = %s, category = %s, publisher = %s, pageCount = %s, whoAdded = %s WHERE id = %s",(bookName,writer,category,publisher,pageCount,activeUserName,id))
+            self.cursor.execute("UPDATE books SET bookName = %s, writer = %s, category = %s, publisher = %s, pageCount = %s, whoAdded = %s WHERE id = %s",(bookName.strip(),writer.strip(),category.strip(),publisher.strip(),pageCount,activeUserName,id))
             self.conn.commit()
                                             
             return {"success":True,"message":"Kitap güncellendi."}
 
         except Exception as e:
 
+            self.conn.rollback()
             writeLog(config.BOOKS_LOG_PATH,type(e).__name__,str(e))
             return {"success":False,"message":"Bir hata oluştu!"}
